requirements.txt
django==1.7
djangorestframework==2.4.2
django-cors-headers
